added genre to be fetched from backend, broke submittion into smaller components
