+++ src/services/index.js
@@ -98,3 +98,4 @@ export const alb_service = new BaseService("v1/audience-leaderboard/")
 export const flb_service = new BaseService("v1/filmmaker-leaderboard/")
 export const contest_service = new BaseService("v1/contest/")
 export const movies_by_service = new BaseService("v1/movies-by/")
+export const genre_service = new BaseService("v1/genre/")
